Add SW Constructors to global scope
