+++ packages/service-worker-mock/index.js
@@ -1,10 +1,22 @@
 const url = require('url');
+
+const Cache = require('./models/Cache');
 const CacheStorage = require('./models/CacheStorage');
+const Client = require('./models/Client');
 const Clients = require('./models/Clients');
-const ServiceWorkerRegistration = require('./models/ServiceWorkerRegistration');
-const handleEvents = require('./utils/events').handleEvents;
+const Event = require('./models/Event');
+const FetchEvent = require('./models/FetchEvent');
+const Headers = require('./models/Headers');
+const Notification = require('./models/Notification');
+const NotificationEvent = require('./models/NotificationEvent');
+const PushEvent = require('./models/PushEvent');
+const PushManager = require('./models/PushManager');
+const PushSubscription = require('./models/PushSubscription');
 const Request = require('./models/Request');
 const Response = require('./models/Response');
+const ServiceWorkerRegistration = require('./models/ServiceWorkerRegistration');
+
+const eventHandler = require('./utils/eventHandler');
 
 class ServiceWorkerGlobalScope {
   constructor() {
@@ -14,10 +26,22 @@ class ServiceWorkerGlobalScope {
     this.caches = new CacheStorage();
     this.clients = new Clients();
     this.registration = new ServiceWorkerRegistration();
+
+    // Constructors
+    this.Cache = Cache;
+    this.Client = Client;
+    this.Event = Event;
+    this.FetchEvent = FetchEvent;
+    this.Headers = Headers;
+    this.Notification = Notification;
+    this.NotificationEvent = NotificationEvent;
+    this.PushEvent = PushEvent;
+    this.PushManager = PushManager;
+    this.PushSubscription = PushSubscription;
     this.Request = Request;
     this.Response = Response;
-    this.URL = url.URL || url.parse;
     this.ServiceWorkerGlobalScope = ServiceWorkerGlobalScope;
+    this.URL = url.URL || url.parse;
 
     // Instance variable to avoid issues with `this`
     this.addEventListener = (name, callback) => {
@@ -30,7 +54,7 @@ class ServiceWorkerGlobalScope {
     // Instance variable to avoid issues with `this`
     this.trigger = (name, args) => {
       if (this.listeners[name]) {
-        return handleEvents(name, args, this.listeners[name]);
+        return eventHandler(name, args, this.listeners[name]);
       }
       return Promise.resolve();
     };
