+++ packages/service-worker-mock/models/Event.js
@@ -1,0 +1,12 @@
+class Event {
+  constructor() {
+    this.promise = null;
+    this.response = null;
+  }
+
+  waitUntil(promise) {
+    this.promise = promise;
+  }
+}
+
+module.exports = Event;
