+++ packages/service-worker-mock/models/PushEvent.js
@@ -1,0 +1,10 @@
+const Event = require('./Event');
+
+class PushEvent extends Event {
+  constructor(args) {
+    super();
+    Object.assign(this, args);
+  }
+}
+
+module.exports = PushEvent;
